Makefile debug rules for callgrind -Wconversion flag artistry

+++ fillit_sources/fillit.c
@@ -6,7 +6,7 @@
 /*   By: thakala <thakala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/11 15:55:48 by thakala           #+#    #+#             */
-/*   Updated: 2022/01/26 16:02:12 by thakala          ###   ########.fr       */
+/*   Updated: 2022/01/26 16:16:23 by thakala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,16 +58,16 @@ int	main(int argc, char **argv)
 	static t_tetri	tetriminoes[27];
 
 	if (argc != 2)
-		return (errors ("usage: Insert one argument after ./fillit\n", -1));
+		return ((int)errors ("usage: Insert one argument after ./fillit\n", -1));
 	fd = open(argv[1], O_RDONLY);
 	if (fd == -1)
-		return (errors ("open_error\n", 2));
+		return ((int)errors ("open_error\n", 2));
 	tetrimino_reference_init();
 	count = 0;
 	if (get_tetriminoes(fd, tetriminoes, &count) <= 0)
-		return (errors ("error\n", -1));
+		return ((int)errors ("error\n", -1));
 	if (close (fd) == -1)
-		return (errors ("close_error\n", 2));
+		return ((int)errors ("close_error\n", 2));
 //function_loader_for_debugging();
 //	system("leaks a.out");
 	return (fillit(tetriminoes, count));
